+++ Radar/src/radar/GestordeMultas.java
@@ -1,3 +1,5 @@
+package radar;
+
 
 import java.util.ArrayList;
 import radar.Radar;
@@ -56,5 +58,22 @@ public class GestordeMultas {
         } 
         tabla.remove(tabla.contains(reg));
     }
+    
+     /**
+     * @param args the command line arguments
+     */
+    public static void main(String[] args) {
+        // TODO code application logic here
+        
+        Radar entrada= new Radar("IB 0198 FWB",20.00, 100);
+         Radar salida= new Radar("IB 0198 FWB",20.20, 125);
+        entradaAutopista(entrada);
+        salidaAutopista(salida);
+        
+    }
+    
+    
+    
+    
 
 }
