+++ Radar/src/radar/Radar.java
@@ -32,17 +32,6 @@ private double kilometro;
         return kilometro;
     }
 
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String[] args) {
-        // TODO code application logic here
-        
-        Radar entrada= new Radar("IB 0198 FWB",20.00, 100);
-         Radar salida= new Radar("IB 0198 FWB",20.20, 125);
-        entradaAutopista(entrada);
-        salidaAutopista(salida);
-        
-    }
+    
     
 }
